Update LaravelIntegration.php
The default horizon work command is `php artisan horizon`. Which doesn't get recognised right now.

I've chosen to do an exact match, because using `strpos()` will match on all Horizon subcommands.

I'm not familiar with the codebase, but I did some quick testing on the `$artisanCommand` variable locally and saw that it contains the command without any options. So this should work I think. I'm not sure why we aren't using an exact match (or `in_array()`) for all options. Is this speed related?